migration and refactor stage 1
