+++ src/App.js
@@ -1,25 +1,7 @@
-import logo from './logo.svg';
-import './App.css';
-
+import AppRoutes from './AppRoutes';
+import "./styles.css"
 function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
+  return (<AppRoutes/>);
 }
 
 export default App;
