+++ src/components/Elements/ElementInputForm.js
@@ -1,0 +1,32 @@
+import { useState, useEffect } from "react";
+const InputField = ({name, validification, inputValue, displayName, type, onInputValueChange, styling}) => {
+    //const [value, setValue] = useState({type==="number" ? 0 : ''});
+    const [validationMessage, setValidationMessage] = useState('');
+
+
+    useEffect(() => {
+      let timeout = setTimeout(() => {
+        let message = validification(inputValue[name])
+        setValidationMessage(message)
+      }, 500)
+      return () => {
+        clearTimeout(timeout)
+      }
+    }, [inputValue[name]])
+
+    return (
+      <div className={styling}>
+        <label>{displayName}</label>
+        <input type={type}
+          name={name}
+          value={inputValue[name]} 
+          onChange={(e)=>{
+          let val = type==="number" ? parseInt(e.target.value) : e.target.value;
+          onInputValueChange(e.target.name, val)
+        }}></input>
+        <label className='input-error-message'>{validationMessage || ''}</label>
+
+    </div>);
+  }
+
+export default InputField
