+++ src/components/screens/Homepage.js
@@ -1,0 +1,8 @@
+const Homepage = (props) =>{
+    return (<div className='Homepage'>
+      <h1>Molecule Viewer</h1>
+      <h2>Daniel Wang</h2>
+    </div>)
+  }
+
+export default Homepage;
